+++ src/core/core.go
@@ -2,8 +2,14 @@ package core
 
 import (
 	"context"
+	"sm-box/src/core/components/configurator"
 	"sm-box/src/core/components/logger"
+	"sm-box/src/core/components/tracer"
 	"sync"
+)
+
+const (
+	loggerInitiator = "core"
 )
 
 var (
@@ -33,24 +39,79 @@ type Core interface {
 //   - StateBooted - "Booted"
 //   - StateServed - "Served"
 //   - StateOff    - "Off"
-func New() (c Core, err error) {
+func New() (cr Core, err error) {
+	// tracer
+	{
+		var trc = tracer.New(tracer.LevelMain, tracer.LevelCore)
+
+		trc.FunctionCall()
+		trc.Error(err).FunctionCallFinished(cr)
+	}
+
+	var created bool
+
 	once.Do(func() {
 		var instance_ = &core{
 			ctx: context.Background(),
 		}
 
-		if instance_.conf, err = BuildConfig(); err != nil {
-			return
+		// Конфигурация
+		{
+			var conf = new(Config)
+
+			var c configurator.Configurator[*Config]
+
+			if c, err = configurator.New[*Config](conf); err != nil {
+				return
+			} else if err = c.Private().Profile(confProfile).Read(); err != nil {
+				return
+			}
+
+			instance_.conf = conf
+		}
+
+		// Компоненты
+		{
+			instance_.components = new(components)
+
+			// Logger
+			{
+				if instance_.components.logger, err = logger.New(loggerInitiator); err != nil {
+					return
+				}
+			}
+		}
+
+		// Состояние
+		{
+			instance_.state = &stateNew{
+				components: instance_.components,
+				ctx:        instance_.ctx,
+				conf:       instance_.conf,
+			}
 		}
 
 		instance = instance_
+
+		instance.Components().Logger().Info().
+			Text("The system core instance has been created. ").
+			Field("state", instance_.State()).
+			Field("config", instance_.conf).Write()
+
+		created = true
 	})
 
 	if err != nil {
 		return
 	}
 
-	c = instance
+	cr = instance
+
+	if !created {
+		instance.Components().Logger().Info().
+			Text("The system core instance was received. ").
+			Field("state", instance.State()).Write()
+	}
 
 	return
 }