+++ src/core/env/env_test.go
@@ -1,8 +1,0 @@
-package env
-
-import (
-	"testing"
-)
-
-func TestEnv(t *testing.T) {
-}