+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # ChangeLog:
 
+### v24.0.4:
+- [x] Доработки компонента [управления конфигурациями проекта](src/core/components/configurator/configurator.go);
+- [x] Доработки компонента [ведения журнала трессировки вызовов](src/core/components/tracer/tracer.go);
+- [x] Разработка [ядра системы](src/core/core.go)
+
+---
+
 ### v24.0.3:
 - [x] Разработан компонент [управления конфигурациями проекта](src/core/components/configurator/configurator.go);